implemented level 3 code
